✨ feat(core): Add Result.collect method to aggregate multiple results
This commit introduces the `Result.collect` method, enabling the aggregation of multiple `Result` objects into a single `Result`. If any of the input `Result` objects are failures, the resulting `Result` will also be a failure, returning the first encountered error. Otherwise, it returns a successful `Result` containing an array of the data from the successful results. Also, update package version to 5.0.0-alpha.5.